requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0

# Database
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# CORS headers
django-cors-headers==4.3.1

# Environment variables (optional but recommended)
python-decouple==3.8

# Production server (optional)
gunicorn==21.2.0
